+++ src/Endpoint/AccountEmailSendersEndpoint.php
@@ -11,7 +11,7 @@ use DigitalCz\DigiSign\Resource\ListResource;
 
 /**
  * @extends ResourceEndpoint<AccountEmailSender>
- * @method ListResource<AccountEmailSender> list(array $query = [])
+ * @method ListResource<AccountEmailSender> list(array $query = [], bool $actions = true, bool $links = true)
  * @method AccountEmailSender get(string $id)
  */
 class AccountEmailSendersEndpoint extends ResourceEndpoint